requirements.txt
asttokens==2.1.0
backcall==0.2.0
category-encoders==2.5.1.post0
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
imbalanced-learn==0.9.1
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
joblib==1.0.1
jupyter_client==7.4.7
jupyter_core==5.0.0
kaleido==0.2.1
kiwisolver==1.4.4
lightgbm==3.3.3
llvmlite==0.38.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numba==0.55.2
numpy==1.21.6
packaging==21.3
pandas==1.4.4
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
plotly==5.11.0
pmdarima==2.0.1
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
pycaret-ts-alpha==3.0.0.dev1649017462
Pygments==2.13.0
pyod==1.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
scikit-learn==1.1.3
scikit-plot==0.3.7
scipy==1.7.3
six==1.16.0
sktime==0.10.1
stack-data==0.6.1
statsmodels==0.13.5
tbats==1.1.1
tenacity==8.1.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
wrapt==1.14.1
yellowbrick==1.5
